'user-replacer'

+++ routes/user/delete.js
@@ -1,4 +1,6 @@
+import { USER_ROLES } from "../../globals.js";
 import { deleteAsset } from "../../helpers/assets.js";
+import { AppError } from "../../lib/error.js";
 import { ArticleModel } from "../../models/article.js";
 import { CommentModel } from "../../models/comment.js";
 import { UserModel } from "../../models/user.js";
@@ -6,6 +8,10 @@ import { UserModel } from "../../models/user.js";
 export const deleteUser = async (req, res) => {
   // create user model instance
   const currentUser = new UserModel(req.user);
+
+  if (currentUser.role === USER_ROLES.admin) {
+    throw new AppError(400, "Admin cannot delete their account");
+  }
 
   // delete user comments
   await CommentModel.deleteMany({
